requirements.txt
# Core data
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# ML Models
xgboost>=2.0.0
lightgbm>=4.0.0

# Explainability
shap>=0.44.0

# Dashboard
dash>=2.14.0
dash-bootstrap-components>=1.5.0
plotly>=5.18.0

# LLM integration
openai>=1.0.0
anthropic>=0.20.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
joblib>=1.3.0

# Evaluation & metrics
scipy>=1.11.0
imbalanced-learn>=0.11.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
jupyter>=1.0.0
